Fix: region skill not expiring (#558)

+++ Maple2.Server.Game/Model/Field/Entity/FieldSkill.cs
@@ -209,7 +209,7 @@ public class FieldSkill : FieldEntity<SkillMetadata> {
     end:
         if (Interval == 0) {
             FireCount = 0;
-            NextTick = int.MaxValue;
+            NextTick = long.MaxValue;
         } else {
             FireCount--;
             NextTick += Interval;
